refactor: Update `study` module structure.
Prepare `study` module for session and timer systems.
Move regular sync to the studybadge loop.
