+++ bot/modules/study/__init__.py
@@ -1,9 +1,8 @@
 from .module import module
 
-from . import data
-from . import admin
-from . import badge_tracker
-from . import time_tracker
+from . import badges
+from . import timers
+from . import tracking
+
 from . import top_cmd
-from . import studybadge_cmd
 from . import stats_cmd
