+++ bot/modules/study/badges/__init__.py
@@ -1,0 +1,2 @@
+from . import badge_tracker
+from . import studybadge_cmd
