+++ bot/modules/study/tracking/__init__.py
@@ -1,0 +1,4 @@
+from . import data
+from . import settings
+from . import time_tracker
+from . import session_tracker
