+++ bot/modules/study/tracking/data.py
@@ -1,0 +1,3 @@
+from data import Table
+
+untracked_channels = Table('untracked_channels')
