+++ bot/modules/study/tracking/settings.py
